+++ common/lib/capa/capa/tests/test_html_render.py
@@ -153,9 +153,7 @@ class CapaHtmlRenderTest(unittest.TestCase):
         # the solution
         expected_textline_context = {
             'STATIC_URL': '/dummy-static/',
-            'status': 'unsubmitted',
-            'status_class': 'unanswered',
-            'status_display': u'unanswered',
+            'status': the_system.STATUS_CLASS('unsubmitted'),
             'label': '',
             'value': '',
             'preprocessor': None,