+++ lms/djangoapps/verify_student/views.py
@@ -256,7 +256,7 @@ def results_callback(request):
 
     # If this is a reverification, log an event
     if attempt.window:
-        course_id = window.course_id
+        course_id = attempt.window.course_id
         course = course_from_id(course_id)
         course_enrollment = CourseEnrollment.get_or_create_enrollment(attempt.user, course_id)
         course_enrollment.emit_event(EVENT_NAME_USER_REVERIFICATION_REVIEWED_BY_SOFTWARESECURE)