+++ docs/en_us/course_authors/source/change_log.rst
@@ -3,6 +3,20 @@ Change Log
 ############
 
 ************
+May, 2014
+************
+
+.. list-table::
+   :widths: 10 70
+   :header-rows: 1
+
+   * - Date
+     - Change
+   * - 05/02/14
+     - Updated :ref:`Drag and Drop` information
+
+
+************
 April, 2014
 ************
 
@@ -12,6 +26,10 @@ April, 2014
 
    * - Date
      - Change
+   * - 05/02/14
+     - Update :ref:`IFrame` documentation
+   * - 04/28/14
+     - Updated :ref:`Show or Hide the Course Wiki Page` to include note about Wiki content being available after you hide the Wiki page.
    * - 04/26/14
      - Updated label information; added XML information to :ref:`Problem with Adaptive Hint`  
    * - 04/24/14